Move `front` role to a separate extension
